Initial import. Copy of HansMeyerXBMC's SuperTV plugin.
This is a copy of code from SuperTV.bundle. The code was published sans
license, and is stored here as a copy. The original code can be found
at http://supercloudtv.com/download.php

"To view in-depth EPG for any channel, press i or hold Menu on the
Apple Remote.
To change the aspect ratio of the stream, press 1.
To change the channel, press Up or Down to reveal the channel list.
To change the volume, press Left or Right.
To view additional options, press c or hold Menu on the Apple Remote"

I have some expectation of value for this, once it works.